[ADD] Add condition dialog
