+++ src/layouts/Condition/index.jsx
@@ -6,7 +6,6 @@ import CardHeader from '@material-ui/core/CardHeader';
 import CardContent from '@material-ui/core/CardContent';
 import CardActions from '@material-ui/core/CardActions';
 import IconButton from '@material-ui/core/IconButton';
-import Typography from '@material-ui/core/Typography';
 import MoreVertIcon from '@material-ui/icons/MoreVert';
 import Button from '@material-ui/core/Button';
 import purple from '@material-ui/core/colors/purple';
@@ -16,14 +15,6 @@ import ListItem from '@material-ui/core/ListItem';
 import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction';
 import ListItemText from '@material-ui/core/ListItemText';
 import EditIcon from '@material-ui/icons/Edit';
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogContentText from '@material-ui/core/DialogContentText';
-import DialogTitle from '@material-ui/core/DialogTitle';
-import TextField from '@material-ui/core/TextField';
-import Checkbox from '@material-ui/core/Checkbox';
-import MenuItem from '@material-ui/core/MenuItem';
 
 const styles = theme => ({
   card: {
@@ -35,6 +26,8 @@ const styles = theme => ({
   },
   actions: {
     display: 'flex',
+    justifyContent: 'center',
+    marginBottom: '10px',
   },
   cssRoot: {
     color: theme.palette.getContrastText(purple[700]),
@@ -47,321 +40,92 @@ const styles = theme => ({
 
 const theme = createMuiTheme({
   palette: {
-    primary: yellow
+    secondary: yellow
   },
 });
 
+const conditionMenu = ['Moving', 'Riding bike', 'Static'];
+const optionMenu = ['weak light', 'lower noise', 'battery', 'speed', 'GPS', 'location'];
+const fromObj = {
+  startHr: 8,
+  startMin: 30,
+  startM: 'AM',
+  endHr: 10,
+  endMin: 30,
+  endM: 'PM',
+}
+const lastObj = {
+  duration: 10,
+  unit: 'minute'
+}
+const conditionObj = {
+  name: 'Condition',
+  schedule_from: fromObj,
+  schedule_last: lastObj,
+  rule: [],
+}
+const defaultCondition = [conditionObj]
 
-const hourInterval = Array(12).fill().map((value, index) => index + 1);
-const minInterval = Array(60).fill().map((value, index) => index + 1);
-const amPm = ['AM', 'PM'];
-const timeUnit = ['hour', 'minute', 'second'];
+class Condition extends React.Component {
 
-
-
-class RecipeReviewCard extends React.Component {
   state = { 
-    open: false,
-    name: 'condition 1',
+    conditionList: defaultCondition.map(c => (
+      
+    )),
   };
 
   handleClickOpen = () => {
     this.setState({ open: true });
-    //console.log("click");
   };
 
   handleClose = () => {
     this.setState({ open: false });
   };
 
-  handleChange = name => event => {
-    this.setState({
-      [name]: event.target.value,
-    });
-  };
-
-
   render() {
     const { classes } = this.props;
 
     return (
       <div>
+      <MuiThemeProvider theme={theme}>
         <Card className={classes.card}>
           <CardHeader
-            action={
-              <IconButton>
-                <MoreVertIcon />
-              </IconButton>
-            }
-            title="Customize"
+            title="Condition"
           />
           <CardContent>
           <List>
-            {[0, 1, 2, 3].map(value => (
-              <ListItem
-                divider = {true}
-                disableGutters = {true}
-                dense = {false}
-              >
-                <ListItemText primary={`Condition ${value + 1}`} />
+            {this.state.conditionList.map(condition => (
+              <ListItem divider disableGutters>
+                <ListItemText primary={condition}/>
                 <ListItemSecondaryAction>
-                  <IconButton aria-label="Comments" onClick={this.handleClickOpen}>
+                  <IconButton onClick={this.handleClickOpen}>
                     <EditIcon />
                   </IconButton>
                 </ListItemSecondaryAction>
               </ListItem>
             ))}
-        </List>
+          </List>
           </CardContent>
-          <CardActions className={classes.actions} disableActionSpacing>
-            <MuiThemeProvider theme={theme}>
-              <Button 
-                variant="contained" 
-                color="primary" 
-                className={classes.margin}
-                onClick={this.handleClickOpen}
-              >
-                + ADD
-              </Button>
-            </MuiThemeProvider>
+          <CardActions className={classes.actions}>
+            <Button 
+              variant="contained" 
+              color="secondary" 
+              className={classes.margin}
+              onClick={this.handleClickOpen}
+            >
+              + ADD
+            </Button>
           </CardActions>
         </Card>
-        <Dialog
-          open={this.state.open}
-          onClose={this.handleClose}
-          aria-labelledby="alert-dialog-title"
-          aria-describedby="alert-dialog-description"
-        >
-          <DialogTitle id="alert-dialog-title">
-            <TextField
-              id="name"
-              label="Name"
-              className={classes.textField}
-              value={this.state.name}
-              onChange={this.handleChange('name')}
-              margin="normal"
-            />
-          </DialogTitle>
-          <DialogContent>
-            <List>
-              <ListItem
-                role={undefined}
-                dense
-                className={classes.listItem}
-              >
-                <Checkbox
-                  tabIndex={-1}
-                  disableRipple
-                />
-                <ListItemText primary='From' />
-                <TextField
-                  id="select-currency"
-                  select
-                  className={classes.textField}
-                  value={this.state.currency}
-                  onChange={this.handleChange('currency')}
-                  SelectProps={{
-                    MenuProps: {
-                      className: classes.menu,
-                    },
-                  }}
-                  margin="normal"
-                >
-                  {hourInterval.map(option => (
-                    <MenuItem key={option} value={option}>
-                      {option}
-                    </MenuItem>
-                  ))}
-                </TextField>
-                <ListItemText primary='：' />
-                <TextField
-                  id="select-currency"
-                  select
-                  className={classes.textField}
-                  value={this.state.currency}
-                  onChange={this.handleChange('currency')}
-                  SelectProps={{
-                    MenuProps: {
-                      className: classes.menu,
-                    },
-                  }}
-                  margin="normal"
-                >
-                  {minInterval.map(option => (
-                    <MenuItem key={option} value={option}>
-                      {option}
-                    </MenuItem>
-                  ))}
-                </TextField>
-                <ListItemText primary=' ' />
-                <TextField
-                  id="select-currency"
-                  select
-                  className={classes.textField}
-                  value={this.state.currency}
-                  onChange={this.handleChange('currency')}
-                  SelectProps={{
-                    MenuProps: {
-                      className: classes.menu,
-                    },
-                  }}
-                  margin="normal"
-                >
-                  {amPm.map(option => (
-                    <MenuItem key={option} value={option}>
-                      {option}
-                    </MenuItem>
-                  ))}
-                </TextField>
-                <ListItemText primary='to' />
-                <TextField
-                  id="select-currency"
-                  select
-                  className={classes.textField}
-                  value={this.state.currency}
-                  onChange={this.handleChange('currency')}
-                  SelectProps={{
-                    MenuProps: {
-                      className: classes.menu,
-                    },
-                  }}
-                  margin="normal"
-                >
-                  {hourInterval.map(option => (
-                    <MenuItem key={option} value={option}>
-                      {option}
-                    </MenuItem>
-                  ))}
-                </TextField>
-                <ListItemText primary='：' />
-                <TextField
-                  id="select-currency"
-                  select
-                  className={classes.textField}
-                  value={this.state.currency}
-                  onChange={this.handleChange('currency')}
-                  SelectProps={{
-                    MenuProps: {
-                      className: classes.menu,
-                    },
-                  }}
-                  margin="normal"
-                >
-                  {minInterval.map(option => (
-                    <MenuItem key={option} value={option}>
-                      {option}
-                    </MenuItem>
-                  ))}
-                </TextField>
-                <ListItemText primary=' ' />
-                <TextField
-                  id="select-currency"
-                  select
-                  className={classes.textField}
-                  value={this.state.currency}
-                  onChange={this.handleChange('currency')}
-                  SelectProps={{
-                    MenuProps: {
-                      className: classes.menu,
-                    },
-                  }}
-                  margin="normal"
-                >
-                  {amPm.map(option => (
-                    <MenuItem key={option} value={option}>
-                      {option}
-                    </MenuItem>
-                  ))}
-                </TextField>
-              </ListItem>
-              <ListItem
-                role={undefined}
-                dense
-                className={classes.listItem}
-              >
-                <Checkbox
-                  tabIndex={-1}
-                  disableRipple
-                />
-                <ListItemText primary='Last for' />
-                <TextField
-                  id="select-currency"
-                  select
-                  className={classes.textField}
-                  value={this.state.currency}
-                  onChange={this.handleChange('currency')}
-                  SelectProps={{
-                    MenuProps: {
-                      className: classes.menu,
-                    },
-                  }}
-                  margin="normal"
-                >
-                  {hourInterval.map(option => (
-                    <MenuItem key={option} value={option}>
-                      {option}
-                    </MenuItem>
-                  ))}
-                </TextField>
-                <ListItemText primary='：' />
-                <TextField
-                  id="select-currency"
-                  select
-                  className={classes.textField}
-                  value={this.state.currency}
-                  onChange={this.handleChange('currency')}
-                  SelectProps={{
-                    MenuProps: {
-                      className: classes.menu,
-                    },
-                  }}
-                  margin="normal"
-                >
-                  {minInterval.map(option => (
-                    <MenuItem key={option} value={option}>
-                      {option}
-                    </MenuItem>
-                  ))}
-                </TextField>
-                <ListItemText primary=' ' />
-                <TextField
-                  id="select-currency"
-                  select
-                  className={classes.textField}
-                  value={this.state.currency}
-                  onChange={this.handleChange('currency')}
-                  SelectProps={{
-                    MenuProps: {
-                      className: classes.menu,
-                    },
-                  }}
-                  margin="normal"
-                >
-                  {timeUnit.map(option => (
-                    <MenuItem key={option} value={option}>
-                      {option}
-                    </MenuItem>
-                  ))}
-                </TextField>
-              </ListItem>
-            </List>
-          </DialogContent>
-          <DialogActions>
-            <Button onClick={this.handleClose} color="primary">
-              cancel
-            </Button>
-            <Button onClick={this.handleClose} color="primary" autoFocus>
-              done
-            </Button>
-          </DialogActions>
-        </Dialog>
+        
+        </MuiThemeProvider>
       </div>
     );
   }
 }
 
-RecipeReviewCard.propTypes = {
+Condition.propTypes = {
   classes: PropTypes.object.isRequired,
 };
 
-export default withStyles(styles)(RecipeReviewCard);
+export default withStyles(styles)(Condition);
